+++ programs/store/src/instructions/gt.rs
@@ -427,3 +427,42 @@ impl<'info> internal::Authentication<'info> for CloseGtExchange<'info> {
         &self.store
     }
 }
+
+/// The accounts definition for
+/// [`update_gt_cumulative_inv_cost_factor`](crate::update_gt_cumulative_inv_cost_factor) instruction.
+#[derive(Accounts)]
+pub struct UpdateGtCumulativeInvCostFactor<'info> {
+    /// The authority for this instruction.
+    pub authority: Signer<'info>,
+    /// The store account to update.
+    #[account(
+        mut,
+        constraint = store.load()?.gt().is_initialized() @ CoreError::PreconditionsAreNotMet,
+    )]
+    pub store: AccountLoader<'info, Store>,
+}
+
+impl UpdateGtCumulativeInvCostFactor<'_> {
+    /// Update the cumulative inverse cost factor.
+    ///
+    /// # CHECK
+    /// - Only GT_CONTROLLER is allowed to invoke.
+    pub(crate) fn invoke_unchecked(ctx: Context<Self>) -> Result<u128> {
+        let mut store = ctx.accounts.store.load_mut()?;
+        let gt = store.gt_mut();
+
+        gt.update_cumulative_inv_cost_factor()?;
+
+        Ok(gt.cumulative_inv_cost_factor())
+    }
+}
+
+impl<'info> internal::Authentication<'info> for UpdateGtCumulativeInvCostFactor<'info> {
+    fn authority(&self) -> &Signer<'info> {
+        &self.authority
+    }
+
+    fn store(&self) -> &AccountLoader<'info, Store> {
+        &self.store
+    }
+}