+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- programs(store): Added `cumulative_inv_cost_factor` as a new global metric for GT, along with an instruction to update it.
+
 ## [0.7.1] - 2025-08-15
 
 ### Breaking Changes